Use views to generate Solr representations.
We no longer have custom document classes, so Solr representations have
to be generated at the database level.  Might as well put them in the
design document; that way, we can use methods already in
Couchdb::Database to generate Solr representations.

There is a problem with the Solr representation of the event: anyone can
name an event anything, but the Solr representation of the event uses
the name of the event as the document ID.  This will cause obvious data
integrity issues.

The current choice of Solr representation for events is the simplest
choice that will satisfy the current autocompletion scenario for events.
There may be ways to fix this without changing the specifications, but
it's also possible that the scenario is too simple.