Couchdb::Database#head.  #23.
Useful for detecting the presence of a document, which in turn is useful
for attaching objects to documents.  (In order to attach objects to
documents, a rev must be provided.  The easiest way to get that rev, if
it isn't already present, is to do a HEAD on the document ID.)