Rework ChangeWatcher using Celluloid::IO.  #22.
The primary motivation for this was to be able to use Celluloid's
supervisors to ensure that a ChangeWatcher stays active.

I'm not sure how to make Celluloid::IO's reactor monitor multiple
objects for readability.  (From what I can tell from its source code,
it's not possible with the stock code -- you could probably reuse
Celluloid::IO::Mailbox, but everything else would have to be coded by
hand.)

Considering that #stop really only exists for convenience of
development and testing, the simpler "set a stop flag and wait for the
thread to stop" method works just fine for me.  The heartbeat has been
sped up to 10 seconds to speed up thread termination.  I'll find out
soon enough whether that's too much of a cost in bandwidth.  I doubt it
is, though:

    1 beat / 10 sec = 8640 beats / day = 8.640 kB / day
      = 259.2 kB / month in heartbeat transfers
   (assuming 30 days / month)

One beat every 30 seconds requires one-third the transfer, but piddling
over kilobytes seems like a silly thing to do.

---

This commit eliminates Couchdb::Database#changes: the way in which it
started a ChangeWatcher made it too difficult to e.g. add an actor to a
supervision tree.  A future commit might bring back a convenience method
for construction ChangeWatchers (Couchdb::Database#change_watcher,
maybe) but I don't see a use for something like that now.