Implement Database#all_docs.
This is useful to have in conjunction with #bulk_docs / #bulk_docs! when
updating many design documents from an out-of-database representation
(i.e. source control).  In a case like that, you know what the design
document should look like, but you don't have the current revs for the
documents.  GET /_all_docs is an efficient way to get those revs.