First cut at view streaming spec.  #3.
The "returns rows as soon as possible" spec is a bit janky, but it's the
best difference I could come up with between parse-everything and
streaming.

Originally, I wanted to test for memory usage, but doing that via
GC.stat is extremely unreliable (not to mention that one needs to code a
different check for each Ruby implementation), and I don't know of any
reliable cross-Ruby environment memory usage check or reference counter.

We'll see if this spec ends up okay.