Make sure that the fiber always returns strings.
The final return from Fiber#resume (or #transfer) is the return value of
the last expression of the fiber.  In our case, it was the HTTPResponse
returned from the request, which we don't want.  This is a bit of a
hack, but works well enough.

I did work out a way to do this with Fiber#transfer, but the only
advantage it had over the present algorithm is that there was no need to
return a zero-byte string at the end.  I'm not sure if that's a good
enough gain to justify it.