Shift, not pop, rows out of staging.
The reason:

Suppose we have n rows in staging.  Currently, we append each one onto
the end of the array.

    r0
    r1
    ...
    rn

If we pop this, we'll reverse the row order and corrupt the view data.

Why did this work with pop?  It was a series of (un)fortunate
circumstances:

- CouchDB chunks the response with one row per chunk.
- Net::HTTPResponse#read_body yields once per chunk.
- Therefore, there was always one row in staged_rows, which meant that
  it didn't matter whether shift or pop was used.

This commit adjusts the all-rows expectation to match.