Disable WebMock for view streaming tests.
There is something in WebMock 1.8.11 that causes the entire response
body to be read, even if it's in a non-mocked request that has been
passed a block.

Aside from this _not_ properly emulating Net::HTTP's behavior, it also
really messes with JSON::Stream::Parser, which has to handle a few
hundred kB of JSON text at once.  This, in turn, causes our specs to
really slow down.