Do not escape document IDs.  #7.
Previously, Analysand escaped all document IDs.  However, for document
IDs that were already escaped, this would double-escape those IDs,
therefore making it impossible to access such documents.

This commit gives escape control to the user, which IMO is the simplest
and most predictable option.

This commit also clarifies that one should _not_ escape document IDs
passed to #copy.  The CouchDB documentation on COPY[1] doesn't mention
this, but I have noticed this to be the case on CouchDB 1.5.0 and 1.6.1.

(I'm not sure what happens if you pass a URI as in [2], but in that case
I'd expect escaping to be required.)

[1]: http://docs.couchdb.org/en/1.6.1/api/document/common.html#copying-from-a-specific-revision
[2]: http://docs.couchdb.org/en/1.6.1/api/document/common.html#copying-to-an-existing-document