Disable problematic streaming example.
The intent of the "returns rows as fast as possible" example is to
show that the streaming view doesn't parse everything up front by
demonstrating that it begins to produce output, well, as fast as
possible.

The time threshold in the example was chosen from no measurement
whatsoever, and over time it's become much less reliable.

I like the *idea* of having an example like this, but its current
implementation is no good.  Hence, the xit.  It may be just one of those
things that doesn't slot nicely into RSpec...we'll see.